bugfix for when closing pair is deleted using the DELETE key
added  two more cases for right and delete key presses.  These now pop
the top of the _matchStack. Previously the right key cleared the stack
causing an extra closing character to be inserted in the user
explicitly closed a pair and the delete key did nothing to the stack
which meant a closing keypress was ignored if first deleted and then
reentered.